In examples, "taxonomic" is the same as "english"
Issue #604, Czech translations (continued)

In translations like

```
ví ucho ― Leonotis nepetifolia (literally, “lion's ear”)
```

the translation part starting with "Leonotis" is has its
classification returned as "taxonomic" due to the heuristics
used in classify_desc().

I've been trying to kludge something better here, but for
this specifically the right call to make is to change it so
that if a description is either "english" or "taxonomic",
that counts as English. There is not meaningful distinction
here in the examples when trying to figure out translation
stuff.

The heuristics could be better, which is what I tried to
figure out, but it works fine for now...