extract_examples: handle special hyphens in arguments
If there seems to be a problem with hyphens in
example template arguments, call a utility
function that extracts `parts` string data
in a different way.

If argument 2 (original language example) has
special hyphens, count them and combine the splits
into one.

Same for argument 3/"translation".

This should also be able to add parts for later
special hyphens outside of these two arguments
or outside of the template call itself.