Add taggable strings.
The string "Rare except in very formal contexts;
alternative in würde normally preferred" is problematic
due to it being split on the ";". We don't use
xlat_descs_map in decode_tags, so adding the string
there doesn't work. A true fix would be to find the place
just before using decode_tags to check cell content
and do a check there, but this is just a single case
and it would add precious seconds to our already long
runtimes. So this is a hack: the string itself is
long and unique, both ways, so split it before-hand
into two and just give the second part "" as its
tag contents so that the "alternative" in it will
not trigger.