classify_desc: New parameter for accepted `english` words
If you have stuff, like taxonomic vocabulary, that you want
classify_desc to accept as valid, add them to an `accepted`
tuple or frozenset (converting sets to frozensets seems
faster than sets to tuples).

"x in set" is faster than "x in tuple".

Using a tuple or frozenset instead of set is because the function
is cached and you can't cache a mutable set.