Synch with changes to Template:+obj
Template:+obj's output has been changed (and become
more complicated), but as a stopgap this should work
for a while until more complex stuff emerges in actual
practice.

Remove the literal plus "\+" from the regex, that's been
removed; for the purposes of +obj, I think every output
starts with `[with`, but there might also be `along with`
and `and` in the mix somehow.

Clean the template output value of all the style pseudo-HTML
wikitext elements and other HTML stuff.