Fix raw_gloss added directly to `glosses`
This should really be looked at again.

Previously, we would add 'duplicate' lines into
glosses when the raw_gloss added and the extra-cleaned
gloss handled when `subglosses` (even when there were no
other glosses other than the first) were handled did
not match. So we got a gloss that is a bit more raw
followed by the same gloss, except slightly more
cleaned. This fix is a dumb one that *tries* not
cause other issues by simply checking if we're
adding the first subgloss and replacing the
raw_gloss with it completly... Which begs the
question, why is the raw_gloss added at all?

Aaah, commit text, what a rubber duck you are. I just
got it; the raw_gloss is added before cleaning
as a super-gloss because it is needed in sense_base
before subglosses are handled, so because the
subsequent glosses are cleaned in there,
the raw_gloss was only handled a little bit.