[en] Replace broken regex with .splitlines()
This regex was originally used to separate
items in a list of glosses looking at "#" and "*"
at the start of the line, but at some point I needed
(yes, my fault, mea culpa) to also split on newlines
because we had that and it didn't seem to break stuff...

But the regex, as used here, is a bit broken even for
that. It would split at the start of every line on
an invisible character and that would create empty
items at the very start of the list, creating empty
strings that didn't do anything and threw off the
indexing (as seen in a later [1] -> [0] change that
I have no idea how it slipped through when I wrote this,
this should have rung some alarm bells).

These empty strings were then filtered out by the
list-generation for iacebam-style gloss filtering.

.splitlines() works correctly.