Clean up supergloss (prev. rawgloss)
The supergloss above subglosses used to be only
perfunctorily cleaned, and that created problems
when we wanted to check if the *same* gloss, when
being added to "glosses", was actually `in` "glosses".

Because the cleaned up lines were different, we
would actually duplicate the gloss; one raw, one
cleaned gloss.

This was 'fixed' in a previous commit, but this does
it a bit better; the supergloss is cleaned, using
(unfortunately) duplicated code from later on in the
function, but I couldn't make the whole thing into a
function, it was just too damn messy and interdependent
and had too many side-effects. Just the lines with
cleaning done on the string were copied.

The issue with duplicated gloss lines isn't actually fixed
with cleaning the supergloss (that's just a nice thing
for superglosses added to "glosses"), that's handled
by a check on the index and not string comparison.