Parse gloss sublist generated by template
If a template generates entries like this:

```
'# {{es-verb form-of|alcanzar}}
```
--->
```
'# inflection of alcanzar
'## first/third-person singular present subjunctive
'## third-person singular imperative
```
the whole expanded text was fed badly into the glosses
and rawglosses.

The fix here is to check if there are lines starting with
`#` in the output, and if so, parse the whole thing and
extract the context like we've extracted `subentries`,
`others` and `contents` earlier in the process, and
replace them with these if they're empty.

I think it's pretty safe to rely on `subentries` etc.
to be empty if we encounter this kind of template.