[en] Handle IPA and EnPR template with same code
These two have the same underlying template, so
they take the same arguments and output similar
expanded text.

The function returns a `base_data` object with
'shareable' tags from the `q=` argument of the
template, the first `(parenthesized stuff)` at
the very start of the string, which can apply
to other data on the same line, if any.

The rest are SoundData entries in a list, because
these templates can have several different `/ipa/`
or EnPR formatted strings with their own tags
and qualifiers.