[en] Push tags from earlier to later sound data on same line
In cases like:

```
* (Received Pronunciation) IPA^(key): /ipa/, EnPR: ipa
```

we want to use the earlier high-level tags ("RP" here)
for other entries on the same line.

This is done in two ways: When handling the IPA and EnPR
templates, we extract the contents of the extracted and
tagged `q=` argument, which is the highest and earliest
qualifier parameter, and then return that "base data"
from the function handling it, which is passed into the
list of template sound data as a separate entry.

The other is using the old `tagstext` variable and apply
that with parse_pronunciation_tags().

When we encounter one of these two, replace the previously
active `early_base_data` with it and apply that to stuff later
on in the line.