[en] Tests for en/pronunciation.py
Found a neat way to mock get_page() output, using
`side_effect=` for patch(); if it receives an iterable
(like a list of Page-dicts), it will return them in
order when called.

Testing these templates with this output might get
out-of-synch with reality of the templates change,
but we can't easily test for that with normal local
unittests.