[en] Remove some unused code from page.py
xxyzz has recently implemented some code for extracting
Chinese and Japanese examples, and quotations, which
is used as a bypass in [en] page.py that returns early
if it find anything good. Some of the old code was
still around, and this cleans some of it that should
100% be obsolete, but some, like the ruby-detection stuff
is left as is, in case we come across something that isn't
taken care of by xxyzz's implementation.