[simple] Process POS with gloss on naked line
There are a lot of Simple English Wiktionary articles
where the glosses (usually/always? a single line) are
without lists (`#`): we usually rely on list nodes
for handling these sections.

Fix: refactor the "parse glosses" portion of the recursive
function into its own function. At first I thought I would
have to make a wrapper around the template_fn so that
I could use the code in all the places it needed to be,
but happily it was actually only relevant to this specific
part where we actually handle the gloss item at the lowest
level of recursion. The wrapping thing worked out, but it
was a bit of a hack.

Tests are really annoying because they work. Copied a
few of the normal tests and removed the `#`s at the
start of the gloss line, and one of them failed when
I had forgot to convert raw tags to cooked tags.