[simple] Collect 'head_templates' data
Simple English Wiktionary does not have word head sections
like other wiktionaries have, but instead each entry (right
after the POS heading) starts with a table or two with the
inflection of the word, generated by templates.

Instead of making a new field, we collect the name and
arguments of these templates and add it to the word entry
data under `head_templates`.