[simple] Fix tests
I tried to hunt down why Pydantic models would return
a `{'': ''}` dict if given an empty dict. Was it because
it was a new object, so it wasn't counted as the default?
But why would a key-value pair appear, as strings??

Turns out, I was looking at the wrong test, and the problem
was in my first test where I had `args={"": ""}` so that
I could easily copy-paste everything...