[en] Retry decoding tags with errors if " and " in tag
Removed unused code with `valid_words` from 2020. Didn't
find an elegant way to use it where it was meant to be
used.

I've been hitting my head trying to figure out what
to do in decode_tags1 to get " and " to parse correctly
(that is, ignored it), and finally gave up. Instead,
I've implemented a new kludge in the vein of the "/"-
kludge preceding this, except thankfully simpler.