[en] Filter out LINK nodes that are files from headers
Fixes #910

Files are links, and so if they have an alt-text that alt-text
would pop up in heads because we handle En heads by looking
at the nodes and don't use clean_value.

We might want to consider creating a LinkNode and FileLinkNode class
like we have for TemplateNode so that we can more easily filter
out file links.