[en] Form descriptions: check if a 'form' has too many words
Fixes #983: "present participle Martinize" was interpreted
as tags ["present"] and the form "participle Martinize".

This was because the code at:

```python
if (
   i > 1
   and len(parts[i - 1]) >= 4
   and distw(titleparts, parts[i - 1]) <= 0.4
   ):
       alt_related = related
       alt_tagsets = tagsets
       continue
```

would see that "participle" is very close (exactly 0.4)
in distw() Levenshtein distance to "Martinize" and
accepted that it must be part of the titleparts
list, like a word in a phrase.

This kludge just adds a test to see if titleparts is
shorter than the rest of `parts`, don't do that.

THIS WILL FAIL IN CERTAIN SCENARIOS, like when an
'inflected' form gains more words!