[en] Disable stray debug print; learned a new Python trick
https://stackoverflow.com/q/10742501

This is very helpful for hunting down forgotten prints.

I first tried to use pbd and `break print` to make a
breakpoint on `print()` calls, but that doesn't
work for builtins, but the above Stackoverflow question
had a suggestion that worked out. Just add the class
and the code into wiktwords and reassign `sys.stdout`
in `main()`, and it will at least tell you the
function that is generating the print.