[el] Replace `int()` calls with `normalized_int()
There was one article that used fancy unicode
numerals ("①") that were accepted by the regex
as digits, but which caused an exception when
fed into `int()`, so instead we have to normalize
any digits received from regexes.

This seems simpler than using `unicodedata.digit()`
or `unicodedata.numeric()` because those would require
looping and parsing the string.