[en] Fix issue with bad form-finding
Fixes #1034

THIS IS STILL NOT PERFECT.

The problem here was that in article
https://en.wiktionary.org/wiki/-%DC%98#Assyrian_Neo-Aramaic

There would be a canonical form "-X " with an extra space.

This space appeared there because it was taken from
what seemed like an empty string, but was actually a
combining character from "-ܘܿ • (-ō)", which my terminal
doesn't render here, but there is a character before the
first space here. The "-ܘܿ" part of the head is correctly
taken as a bit with forms, but it is partitioned into two
parts, ["-<NORMAL GLYPH>", "<COMBINING>"].

This is caused by the alternative "links" regex that
adds links into the main regex to detect words, and which
also was used to add problematic article titles with
non-alphanumeric characters.

The issue with this is that the article title was only
"-<NORMAL GLYPH>", not "-<NORMAL><COMBINING>", so the regex
would have "-<NORMAL>" overriding the normal word-finding
regex.

Why was the article title added to `links`? Because it had
a non-alphanumeric character "-".

This cannot be fixed with Python's regex handling!!

THEORETICALLY you could add `\b` to the start and end
of every phrase/word found in `links`, because `\b`
stands for a virtual "word-boundary" character that
only exists when a word (`\w`) and non-word (`\W`)
character are neighbors. So that should work, right?

Combining characters do not count as word-characters!!

"-<NORMAL><COMBINING>" is actually "-<NORMAL>\b<COMBINING>".

The hotfix here is to check if the article title starts
with "-" and not adding it to links if the rest of the
title does not contain non-alphanumeric characters.

Instead of "isalnum()", we might need to check each
separator character instead, but that is resource-intensive.