[en] Add "Nouns" and "Proper Nouns" linkage section titles
See: mouth, heath, town

If we have a section like

```
====Derived terms====
* {{lookfrom|en|town}}

=====Nouns=====
...

=====Proper nouns=====
...
```

fix_subtitle_hierarchy will demote "Derived Terms" to level 6
(due to moving all Pronunciation sections between 3 and 4)
and so `Nouns` and `Proper Nouns` are at the same level and
don't get parsed at all.

Fix: make section titles "nouns" and "proper nouns" into
linkage section names. This seems to be the case in
the database currently, "Nouns" and "Proper nouns" are
not used as POS sections.