[el] "bl"-template handling with template_fn
Work based on tests and some logic in PR #1094

The tests seem good to have, so I worked towards those,
and replaced the logic of the original PR with use of
template_fn stuff to capture "bl"-template data using
closures with subfunctions. This leads to some trivial
code duplication in the template_fn declarations in two
places, but that's not the end of the world.

Instead of modifying the tree, we prefer to intercept
the output of turning that tree into text, using first-class
functions that can replace the output of a template
(or node, if using `node_handler_fn=`). A bit clunky
because the main way to communicate with the outside world
(besides inserting meta-strings into the text itself,
like is done with the bold and italics `__I__` strings
elsewhere in the Greek extractor) is through closures.

We extract "bl"-template ("check out this related [[word]]"
template) data into a list that is added to `related`,
and do some kludgy stuff to make it a special case in
case there's no gloss-text to attach it to.

If there's a "synonym"-template in the same text, put
them in `synonyms` instead. Other templates of this
kind are easy to add if needed.