[el] Add a simple test for when a template has a language code
In these inflectional form-of templates, the second argument
can be an optional language code, which is slotted into
`template_parameters[2]`, using an integer index.

Templates parameters work roughly like this: You have numbered
slots and named slots with string indexes. If there is no `=`
in the argument, then its put into a numbered slot, if there is
then its given a string key, except if it's a number then
its put into a numbered index.