[en] Yield td-elements that are erroneously children of TABLE_CELL
There are tables that mix td-elements and wikitext table syntax.

This is most probably done because it doesn't matter in Wikitext
whether something is a wikitext table or HTML-element, they
all get expanded into HTML eventually.

However, it wreaks havoc on our parse-tree, because in these
cases td-elements are made into the direct children of a
TABLE_HEADER or TABLE_CELL node, when they should be "siblings".

td-elements are mixed into Wikitext tables probably? because
they're easier to write #if-conditions etc. for. Just clearer,
in wikitext.

Instead of "correctly" parsing this (which is contradictory:
there might not be a 'correct' way to do this), we instead
correct for this in code by checking if a `col` node's children
include direct child td-elements, and yield those, too.