[en] Implement and kind of ignore syndiff and synsee in senses
Template:synsee and Template:syndiff (two almost identical templates
used intechangeably, afaict) create a "see:" synonyms entry.

However, often this is a Thesaurus namespace link. Currently
we ignore those for sense synonyms, probably because it would be
really messy, in case that Thesaurus data is somewhere else in
the word data. It doesn't make much sense for "Thesaurus:jó" to
populate a single sense of the word `jó`.