[en] Introduce `hyphenations` field, handle `hyph` template
Instead of having a field `hyphenation` (which should be
deprecated) with a list of strings, this introduces a new
Hyphenation TypeDict with a `parts: list[str]` field for
the different parts of a hyphenation and `tags`.

Currently we will have these two parallel and try to
remove `hyphenation` later.

`hyphenation` and `hyph` templates are also handled now
in the template-handling function in pronunciation.py,
in a similar way to pron and audio file templates.

However, some language-specific templates have their
own modules that generate their own hyphenation text
(looking at you, it-pr...) so this won't handle that.

"Why doesn't this Italian language example work?? Wait,
`quieto` doesn't use `hyph` like the template documentation
example said, it uses its own template..."