[en] Add "attestations" field to senses (Template:defdate)
Issue #1278

Done within the gloss/sense template handler function.

`Template:defdate` is used to show the earliest (or just early)
attested dates in text where a word, or a word sense, has been
found. It's main parameter (#1) is a string data like
"early 19th century", and it can have up to three `refX=` and
`refXn=` arguments. The ref (based on how it's in etymology
templates) is probably a `Template:ref` style reference that
is then output by the Wiktionary expander with a `<references/>`
tag later in the article, while the `refn` arguments refer to these
without repeating the `ref` part. I think, based on a quick look,
that we don't do special reference handling with `ref` template
data, but the same data can be found within the word data so
users should be able to access it.

For example in the article `gratis` for Swedish, there's a `ref`
in the etymology section which is later refered to by `refn`
arguments in later `defdate` templates. Both pieces are available
and the `refn` should be unambiguous.