[en] Sort senses according to their original order
There has to be a better way...

English senses are only added to the list of senses
when it is necessary, that is, usually we don't have
the supergloss with the subgloss. However sometimes
there are things in the supergloss (like examples)
that aren't in subglosses, so we add the supergloss
in addition to the subglosses... But because this is
done recursively and we plop the new entries into a
messy nonlocal list called (renamed) `sense_datas`
the order is depth-first and the supergloss would be
after the subglosses.

This "fixes" things by adding a temporary field to
these senses that is generated by another nonlocal
called `sense_ordinal`, which is copied and then
accumulated at every level of the recursion tree.

Note that it is not data, it is purely internal,
and is deleted immediately after we sort the list
of senses. The value is never reset manually, and
it skips numbers when it is incremented on empty
levels of the recursion. DO NOT USE IT FOR ANYTHING.