[en] Add `column-index` inflection data control flow field
See #1313 part 3

`grc-conj` table for Ancient Greek generates a table where
we would like to reset previous headers to none when we come
to a specific section. However, we couldn't do that previously
because we did not have any place to "attach" the `dummy-reset-
headers` pseudo-tag in the inflectiondata control flow dict.

If we attached it to "active", it would trigger earlier in
the table, because "active" is present earlier; same for
"middle/passive", it's present in two places. If we attach
it to "infinitive", we would wipe out the previous row's
headers, which is not what we want.

This commit introduces a new control flow field for infl_map
called `column-index`, which checks if the control flow currently
is in a specific column (zero-indexed). In the above case,
we can distinguish the first "active" (column 0) from the second
"active" (column 2) and attach a reset tag to the second one.