[en] Revert changes to `add_def`, remove ref-elements in `after`
Reverts PR #1323

Even though there's nothing "wrong" with the actual code changes,
except a minor resource cost because we do more things, if it's
not actually needed I decided to revert it.

The actual "fix" to this seems to be that when parsing the
output of tables, we do some bespoke local recursive delving
of the parse tree for some reason (probably makes sense),
so `clean_node()` is not actually used here to get the `after`
text of the table containing footnotes.

Solution: when recursing, skip `ref`-elements.

The broken ref-less version of the taggable string text
that was problematic was move from `xlat_descs_map` to
normal tags.