[en] Georgian ka-decl-noun kludge extended to comma-separated alts
See #1286#issuecomment-3280297085

The previous kludge assumed this table would only have
one entry per cell of Georgian+romanization, with alternative
forms shown with suffixes in parentheses (suffix and no suffix),
but turns out these can be comma-separated alts.

The comma separated alts are already separated earlier in
the text, so we only need to loop over them (and also check
the length of `alts` with `>= 1` instead of `== 1`...).