[en] Enable mult. poses in form_transformations, fix issue with romanization
See #1313

Make the POS `field` (this should be a dict not a list...)
be optionally a tuple of strings instead of just a string,
to enable multiple POSes (like noun and proper noun).

There was a bug with romanization, where romanizations
were not being run through a `form_transformations`
loop. A bit costly, but if the original language form
has a match and it has a romanization, it almost
always has also an equivalent transformation for the
romanization.