[el] `form_of` missing arguments: take the next links
See #1372

Previously we only extracted `form_of` data directly
from template arguments, but because we cannot get rid
of Wiktionary users they often just use a naked template
without arguments and then add a separate linked text
after the template for some forsaken reason. May they
stab their toes.

This fixes part of the issue by giving the function
access to all the nodes in the same list of nodes,
which can then be used to look for empty text and
links immediately after the naked template. These
are then used as the lemma of the `form_of` field.