[el] Improve head splitting and fixed bugs
Co-authored by daxida

https://en.wikipedia.org/wiki/Help:Colon_trick

Template:ετ was creating [[:Category:...]] links, which
the code didn't handle.

Also fixed bugs with how the return value of
`wxr.wtp.node_to_text` was being unpacked, even though
it was a string... Oops. Well, the contents would have
been concatenated at some point anyhow, so there was no
(afaict) actual changes in behavior and ultimate output.

Also handled HTML tag italics and bolds better (that is,
at all).

Changed the logic when looking at the contents of a link:
if there is bold text in the head, do not accept italics-only
as a form.