Cleanup of sorted lists and lists that should be sorted
Quick check looking for lines where we have both `list(`
and `set(`.

Remove unnecessary `list()` that makes return of `sorted()` a list
(which it already is).

Use `sorted()` instead of `list()` when using `set()` to remove duplicates.

Replace `list()` followed by `.sort()` with `sorted()` if
I came across it (not an important change, but might as well)