Move code that modified `valid_tags` and `valid_topic`
See #1427

When `form_descriptions` was moved to the new English
extractor from the top level, it also had some loops
that added stuff to `valid_tags` and `valid_topics` that
more sensibly belonged to the top-level.

This also caused issues because if code from `form_descriptions`
was imported into another extractor, it could cause
side-effects in tests depending: `valid_*` could be
different from what was expected.