[en] When checking linkage sections, recurse through HTML lists
Also see wikitextprocessor PR #401 that adds `is_list()`
and `is_list_item()` functions.

Part of the issue in #1398 was that `Template:col` was not
pre-expanded (all the other aliases of col were, though??),
but even when pre-expanding there was an issue because the
expanded template outputted HTML lists instead of wikitext.

It is correct to also recurse through HTML lists and list
items, and fixing it did separate the different lines so that
the output was not "concatenatedtogetherlikethis", but
the way linkage items are parsed assumes that the hierarchy
is like this:

* something that should be a tag:
** example forms

NOT:

* an example form
** an example form with a suffix
*** a third example form derived from the item directly above

This needs checking.

Do not merge until the PR in wikitextprocessor has been merged.