[en] Start refactoring monster parse_linkage() code
Separate out subfunctions outside and do things
with parameters instead.

Move parse_linkage_item() (which already was separate)
to the end of the file.

Cry.

I've been trying to figure out the best way to do
this for some changes in the structure, and in the
end I think I've come almost full circle back to
"actually let's not change this structure too much".

Actual changes in behavior in later commits...

Before that, I'll try to clean up and remove
unused code and parameters (like `italics=`).