[en] Ignore Template:table auto-filling tables in linkages
I also tried to add middle-dot "·" as a word separator for
linkages, but that doesn't work because it is used in
MORSE CODE strings as the dit... Which means we can't
parse stuff like `Template:list:planets of the Solar System/en`
properly. Even using spaces around the middle dot reads as
Morse code "e".